@@ -1,7 +1,6 @@
-﻿namespace SudokuSolver
+﻿namespace SudokuSolver;
+
+public static class SudokuSolverVersion
 {
-    public static class SudokuSolverVersion
-    {
-        public const string version = "0.4.0-alpha";
-    }
+    public const string version = "0.4.0-alpha";
 }
