@@ -1,33 +1,30 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
+﻿using System.Diagnostics.CodeAnalysis;
 
-namespace SudokuSolverConsole
+namespace SudokuSolverConsole;
+
+public class ByteArrayComparer : IEqualityComparer<byte[]>
 {
-    public class ByteArrayComparer : IEqualityComparer<byte[]>
+    public bool Equals(byte[] x, byte[] y) => x.SequenceEqual(y);
+
+    public int GetHashCode([DisallowNull] byte[] data)
     {
-        public bool Equals(byte[] x, byte[] y) => x.SequenceEqual(y);
-
-        public int GetHashCode([DisallowNull] byte[] data)
+        unchecked
         {
-            unchecked
+            const int p = 16777619;
+            int hash = (int)2166136261;
+
+            for (int i = 0; i < data.Length; i++)
             {
-                const int p = 16777619;
-                int hash = (int)2166136261;
-
-                for (int i = 0; i < data.Length; i++)
-                {
-                    hash = (hash ^ data[i]) * p;
-                }
-
-                hash += hash << 13;
-                hash ^= hash >> 7;
-                hash += hash << 3;
-                hash ^= hash >> 17;
-                hash += hash << 5;
-                return hash;
+                hash = (hash ^ data[i]) * p;
             }
+
+            hash += hash << 13;
+            hash ^= hash >> 7;
+            hash += hash << 3;
+            hash ^= hash >> 17;
+            hash += hash << 5;
+            return hash;
         }
     }
 }
+
