@@ -1,13 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using static SudokuSolver.SolverUtility;
+﻿namespace SudokuSolver.Constraints;
 
-namespace SudokuSolver.Constraints
+[Constraint(DisplayName = "Anti-Knight", ConsoleName = "knight")]
+public class KnightConstraint : ChessConstraint
 {
-    [Constraint(DisplayName = "Anti-Knight", ConsoleName = "knight")]
-    public class KnightConstraint : ChessConstraint
-    {
-        public KnightConstraint(Solver sudokuSolver, string options) : base(sudokuSolver, "1,2") { }
-    }
+    public KnightConstraint(Solver sudokuSolver, string options) : base(sudokuSolver, "1,2") { }
 }
