@@ -1,25 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace SudokuSolver;
 
-namespace SudokuSolver
+public class WrongLengthGivensException : Exception
 {
-    public class WrongLengthGivensException : Exception
+    public WrongLengthGivensException()
     {
-        public WrongLengthGivensException()
-        {
-        }
+    }
 
-        public WrongLengthGivensException(string message)
-            : base(message)
-        {
-        }
+    public WrongLengthGivensException(string message)
+        : base(message)
+    {
+    }
 
-        public WrongLengthGivensException(string message, Exception inner)
-            : base(message, inner)
-        {
-        }
+    public WrongLengthGivensException(string message, Exception inner)
+        : base(message, inner)
+    {
     }
 }
