@@ -1,18 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Xml.Linq;
+﻿namespace SudokuTests;
 
-namespace SudokuTests
+internal static class Puzzles
 {
-    internal static class Puzzles
-    {
-        public static readonly string blankGrid = ".................................................................................";
+    public static readonly string blankGrid = ".................................................................................";
 
-        public static readonly (string, string)[] uniqueClassics = new (string, string)[]
-        {
+    public static readonly (string, string)[] uniqueClassics = new (string, string)[]
+    {
             ("........1....23.4.....452....1.3.....3...4...6..7....8..6.....9.5....62.7.9...1..", "324876591517923846968145273471238965835694712692751438246517389153489627789362154"),
             ("........1.....2.3...4.5.6...3...1...4.78.....6..54.....4....7...9.....127...6.8..", "962437581175682934384159627538791246417826359629543178243918765896375412751264893"),
             ("........1.......2...3..45......6...7..5..8....9.3..4...39..58...8.9.....2...1..7.", "962587341541639728873124596318462957425798163796351482639275814187946235254813679"),
@@ -157,19 +150,19 @@ namespace SudokuTests
             ("..............1.23..2.4..51.......6...1..534..7......5..3..41....481.5..8..9....4", "159326478487591623632748951245137869961285347378469215593674182724813596816952734"),
             ("........1.....234...1.5..62...7.......2...4...8..2..16.15.89....68......2...6..85", "823647591596812347741953862154796238672138459389524716415289673968375124237461985"),
             ("........1.....2.3...413.5.......6..7..5..1..3.3.54.1...18..3.9..43....1.5..8.....", "352694871861752934794138562189326457475981623236547189618473295943265718527819346"),
-        };
+    };
 
-        public static readonly (string, string)[] uniqueClassicFPuzzles = new (string, string)[]
-        {
+    public static readonly (string, string)[] uniqueClassicFPuzzles = new (string, string)[]
+    {
             (@"N4IgzglgXgpiBcBOANCALhNAbO8QCUIAHInMEVAQwFc0ALAewCcEQAFOiLYgAgDkYAdwC2lAHYUQTamRhpWfZqKw8w1ACYMA1tR7SyPSiSwBPAHQ8AgmPU96lNAHIwPTAEJJAcyYR1CANr+oABulFjUuCggnhDBMBLwaNIwAL7IwGkZ6ZmZIKHhuAAcqDFxCUkRufkRCADsJbHxCBWp6XlhNfAAbA1lzck5ALrIQTnZbdW4AIy9TYkDEx2Rs+ULWetDIxuLBQgATCv9lePbpynDozudAMyH88dnJ2MZF+vtu0h3Lc8/J++dM2ijVWD3+uB6QL691SrzBCGKkLm3yeKLekwQAFYvms4Z9ESDWi8trj6vijoTfmilggIaUkTj0fAEXSCVVqfAsWToZsgrjOSzyZS2R9mcDBVdllzkSF2bSxdyJXVsQ9zlshYqOcqKajcQAWLXnQYpIA===", "913548762678192435452637189137965248295874316864213597749326851521489673386751924"),
             (@"N4IgzglgXgpiBcBOANCA5gJwgEwQbT2AF9ljSQA3AQwBsBXOeAZlTQgpgDsEAXDBkqGr1GKdOy69+MQbNJyhtBggAsrCd3h8BAXWSEFh+ccUiEAVnUdN2mfL0Hyw5fABsVyVukLKSxgA4PG28TXzN4NXFrKQFQ50YARiCYu2IHMgywlwB2ZK9YzPiEFijPWyNCvwQxNmj8mXSKrMYAJjzyuKq3dpCM2XTm1R6CpqL4S1Lgkacuktqy3v79SvDAyZSfMdz1+orBQe6djrTl/bGko8WTPftlg5qNDevn0xcL+anUg4mPp76dHREIA=", "823915674741362958569874213657438129138296745492157386385729461216543897974681532"),
             (@"N4IgzglgXgpiBcBOANCAhgVwC4AsD2ATgiAGIA2hEaAdgAQDqhWAtjdSKgRmTGDFsQByhVmVoBlDABM8Aawy0uPMBxABzAhCkIA2juABfZIeNGT5kADc0ZDHHgBWVGoiWY7eFi4wzVm3YQARmdXdwQvO19rW3sAFhC3DwifUwBdZH0o/ziEsM9vLJiEADZcpILTYz8i+ABmMvCK8zNCgPgAdgb8yPTMqui2gA4u5JbKiwH7YPVQ8sjx1vsAJhGC3ub+7IRhmcTG+Y3QSYR43bzRhc2a+rO5n3XFhE7b/ZTD6rbplz3ut7GjrbwFYvX4tdYfZarA7/R4dKF/cYQhBOEEXQzg47wUqopqw07fc5NJF1eEwq5tFA46FpDITQE3Al3Ml0mo7RmvZkGB6XAE1bHs0HknJUhGHLmpAxAA=", "392785146146932587785461392518294673473516829269873451654327918937158264821649735"),
             (@"N4IgzglgXgpiBcBOANCALhNAbO8QDUBDLNQgJwhFUIFc0ALAezIXHoE85UyacwY0rAHbMAtsQAEYGgBNGAaxoSefCYQAO6rOyogA5hRkIA2sdAA3YjVwAWVHojmYQhGh4wAvsmBef33wH+QRZWuABM9o7Oru4eALrIZoEhWNYIAKyRTi7wbtbJIJapuAAcWdG5sd6FoQgR+lE5eZ7VRWlI5U1Vfj4JSa218PUO2TH5wckFbbgAzJ1jLb2JPb41xQhlDaOV4z1r7QCM8zuLk/59Ke2ZWxXNq9MIAGzHdxMD6x03XbtTgwDsL26+1KgPyFzOewe8GeXwWvw+1xGt26AXB70OoNObxW6NwMKR31OaMhg3xjThuIQR1hJ3ug0R5NpwWBCABNNeKwuLPgc3ZKOxELpHzsfLBcQ8QA", "483971652675382941129645837948517326561293478732468519814739265296154783357826194"),
             (@"N4IgzglgXgpiBcBOANCALhNAbO8QGEsBDMSAYwAIBlAVwBMB7AaxooGIAWEVImtACwYAnBCABWDAG4QA+kSzcQQmjjAw0ogHLCAtvIph6zVstUUiABwtYAngDpFAcyEQ6CANrvgAX2ShJ8jS4AGyojhCSMAB2CGjKML4+fomJIAFYQQgoIOGRMfBxQSl+aYG4HGER0bHxKQC6yF6lGbgAjJV5NUXJPUn+ZQgAzB3VBbW9xf0tCACsI/mFCQ1ek6npmfAATPNdCb3NG6E5VQvjfWsD8AAcO2NFy+cl67gA7LeLq/vPWe9n9Y19A7lX7dR5gi7TLYgvY+B6TIEIdrHTp3GGfKYbYbI0YfCYPBHwI65HF/L6XG7Y06giEbCqU3bo7z477wOb01GMp6XN7s3HgrmQ7LEqlLAE03DbXmkjG4IUnBkTMmQinChWwureIA=", "765189243182437695934256178857623914346591827291748356679814532518372469423965781"),
-        };
+    };
 
-        public static readonly (string, string)[] uniqueVariantFPuzzles = new (string, string)[]
-        {
+    public static readonly (string, string)[] uniqueVariantFPuzzles = new (string, string)[]
+    {
             // "Clipped" by glum_hippo: Arrow, Thermo, Givens, King
             (@"N4IgzglgXgpiBcBOANCALhNAbO8QGEsIAHYmAExFQEMBXNACwHsAnBEABQYiNIAIAQlloBbGH2oBranwDmwkQH1upJlRAtaOMDDTsAyrXJNJtPgFo+MAG4wWATz4smAd2R8AxkwUA7d9R9yPgBmAA9gvgAjJlC+EVowNE8mHzRqCB8JLCw+RnFyCFlMMD4AM2cRPgBGc0QAOj59JjE+AqK0EuoWcWosbupyRyLbHzqAHR8JgEFUiHMAaQzZCysAR1pe1sLiuOpHHyYktCZaDwYt6lkU3qx7ccmfABUGOxFm3TsVtp34xL5ElgQDzYRwZDz9HRlCpRLSRXJMXIke7TFjOFwrPKeCAsDw4ILfJJeHyJTD0GAlTFgUR8JilXIvLbtTpYFLLTFdNH3dSyQGUeAAbX5wAAvsgRWLReKpZKZRKALrIIWy6US1UqkUKpVq5Ugay9Wi4ABsqGGMB8CDQmhgyptmvVNu12rtOr1wlwAFYTRARharS79bgql6ffBLQb/W6EAAWYNm33hp2K+2O5MapMOqW6gMIYKx82hv0pmXOoul6Ul1MZ4sKkABDCSJbxmCoPIsN5iNB2BBCkBEHzk7v8kAAJUN+Hd6mHAHZ8FHJwAOfDBBf4ABMIDlcslvYyA4FQ+H7rXk6jS8nwVnk9Xl83277e8FI9PE9Qw4vhqv+Cnn/nG63Yp3fswEHEcj1/V8x2/V8Zw/V9Fw/W8aw5Vxu1Ae9gP3UDL1fU9l1fC911fa9103VAPBgbIMIPI853/NDdyo6j8FgkdIMnGdwJHRdf1IkByMo7ssIQu8GJA4dFxfEcZznCCz1kkia34rAqK48cNxEoCxOvSS32Yk8v30njFIo5TBOHbT1IA9CxJnPDVJkkdEDU3ilJU6czy3LcgA=", "867452931342891675519763284984327156635918427721645398498536712276189543153274869"),
 
@@ -253,6 +246,6 @@ namespace SudokuTests
 
             // "Equal Sums" by Rangsk (Region Sum Lines)
             (@"N4IgzglgXgpiBcBOANCALhNAbO8QFEBHAVwEMsACAZWIFswRVTi0ALAewCcEQAlUgHYBzMAGtGITsRxgYaHgDkutctWIATdqOIUpMiqQAOhrAE8AdAB0B1gGJcKMUgGNWFLBAExkFdRCGYYBSspABuMAYCjiSqYHQUAO6YrJ6OLm4ARuwAHhSYFIakYLJBbJzsxEKs5hJCnBDqCADaTaAB4QIACjACzhBYALKknKIM8ALSWAC+yG0QHd29/UMjYxNY07Mg7T2LfYPDowjrm3MLPfsrR+OTM2e7F8uHa7dbO12PB6vHr/cfS19ric7tt5g8AVcXhsQe89k9vjdoQBdZCtUHnCHPH7Qt5g/6XLGI07o8EEhHA3EYslA34k/Hwmk4v5wwFQ4mwz6Q7HsvEsrlEmG8zmE4EotEczHk2kS6lswVUhlyymkxXc+Uq1lq5X0zUC7V8kXSoWSxk8hW60Wo5nCqVMukG21mjX8inWk1Kt2yrWe1V6n0Wo3ml2B52G5FW+0203qnXBu0y32uyPu73Jr1+tOJkOxsNOnOOmMO6Ni/1xvNFj2ZgPx43ppMJ6vlqOVhtlwvN1Ot3NTEtVtv6jsZrsFgcpoe1rM1oPd0d17MVzsTxs9iPD6OzydNsf1pf90sz/cjw/r48t3fd3trs/To99g8gULkYi4ADsqHeCDQUm8d9vV8XN4nr+QH/uOgFykiKKSDAATsAIcS0B4XjNKASEwGMLR8AAjAAwlhEi8AATHhBEAMw4YRBEACwUdROGkXRVEgJBIJoRhTR8IgJGoLwAAc3F8C+Am8AAbLRPFiQxEk4UxLFbGxzQcbwNEiXRL50bxZE4ZpPHETp2E4YgzFIqxnjoYpfBiap0nqdJ+m8EJ9n8UZPFcUZcmoWZ7GceJPlST5TGuThACsBFcapHkgAp8CYXxOHWT5tk+ZpkXRbF/GBYJMkEUJoXSXlfDBSFxmmV43lxZRPFCZVlnCUVNW8EVDGpV5FlEYZWkuXwNFdY1HUtWVbU0QVvDkQl7URSZ8mtTFSm4ZlvC4SNuHjbhSWLdpJVQX4YCkBkODqFg7ABM4zRQV+z7OIIfjqKQaDoewhgYHBGFQWhKFRTNsW4fhunCeRDU0YD9GMcZqAVNgZk4TwADEhH4DhABCti2BISTqGwCAAAzmIRwWoAAZuUtA4S9X6kJ48h4LwMEQHBai0BQAAyZkgKV5mzT5v18Px3MOcJYkNZJBFibJUEQ2h0N4HDCPI6jqDo5j8A43jhPE6T8Hk5TPA07BUQ0IzLPIez5UqWpGlafZekEbh7niywkuw/DSMo2jDRKyr+MgET7Ak2TnAUwIVN8LT9MG8zrMm21Vkizh61ifZjkEc5YX9fbkNeFLIAyy78sgIrrDY7jXs+37msB9r1Oh/r8RG3AUec7wXENVx/lN9lQUjeFYMgBLUNO7LrsK+7hfK8Xau+xrYBa0HOvVwzEfG9Ng2N/x41cetXEpenjvS87ctuxjo+exPZfTxXs9V3rC912zy8c+lHdZQtuWxyNRWhZB4MO/3e+D3nBci6q29urf2gdg66zpjXQ2kd77lX4g1aqsc+b1QIk1HufdM4D1zofD248QGTzAZXEO19w63wbrFYivVyK9R6mgtO38M4wCzjnA+w8j5AJLqA8u4C56kNrrAzyK9YrDS0uNYik1GG72zvvIe+cR6cNPlPGeED55kMEZ9YRc0n4bWWvFG2ccDHbykb/GR/9cHH3waXZRF9VH8JgcbEyQA=", "146583972593672418728491536917836245632145789854927163465319827371258694289764351"),
-        };
-    }
+    };
 }
+
