@@ -1,79 +1,69 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SudokuSolver;
-using static SudokuSolver.SolverUtility;
+﻿namespace SudokuTests;
 
-namespace SudokuTests
+static internal class TestUtility
 {
-    static internal class TestUtility
+    public static string ToGivenString(this Solver solver)
     {
-        public static string ToGivenString(this Solver solver)
+        var flatBoard = solver.FlatBoard;
+        StringBuilder stringBuilder = new(flatBoard.Length);
+        if (solver.MAX_VALUE <= 9)
         {
-            var flatBoard = solver.FlatBoard;
-            StringBuilder stringBuilder = new(flatBoard.Length);
-            if (solver.MAX_VALUE <= 9)
+            foreach (uint val in flatBoard)
             {
-                foreach (uint val in flatBoard)
-                {
-                    stringBuilder.Append(GetValue(val));
-                }
+                stringBuilder.Append(GetValue(val));
             }
-            else
+        }
+        else
+        {
+            foreach (uint val in flatBoard)
             {
-                foreach (uint val in flatBoard)
+                if (val <= 9)
                 {
-                    if (val <= 9)
-                    {
-                        stringBuilder.Append(0);
-                    }
-                    stringBuilder.Append(GetValue(val));
+                    stringBuilder.Append(0);
                 }
+                stringBuilder.Append(GetValue(val));
             }
-            return stringBuilder.ToString();
         }
+        return stringBuilder.ToString();
+    }
 
-        public static void TestUniqueSolution(this Solver solver, string expectedSolution)
-        {
-            Solver solver1 = solver.Clone();
-            Assert.AreEqual(1u, solver1.CountSolutions(multiThread: true));
+    public static void TestUniqueSolution(this Solver solver, string expectedSolution)
+    {
+        Solver solver1 = solver.Clone();
+        Assert.AreEqual(1u, solver1.CountSolutions(multiThread: true));
 
-            Solver solver2 = solver.Clone();
-            Assert.IsTrue(solver2.FindSolution(multiThread: true));
-            Assert.AreEqual(expectedSolution, solver2.ToGivenString());
+        Solver solver2 = solver.Clone();
+        Assert.IsTrue(solver2.FindSolution(multiThread: true));
+        Assert.AreEqual(expectedSolution, solver2.ToGivenString());
 
-            Solver solver3 = solver.Clone();
-            Assert.IsTrue(solver3.FillRealCandidates(multiThread: true));
-            Assert.IsTrue(solver3.IsComplete);
-            Assert.AreEqual(expectedSolution, solver3.ToGivenString());
-        }
+        Solver solver3 = solver.Clone();
+        Assert.IsTrue(solver3.FillRealCandidates(multiThread: true));
+        Assert.IsTrue(solver3.IsComplete);
+        Assert.AreEqual(expectedSolution, solver3.ToGivenString());
+    }
 
-        public static void TestInvalidSolution(this Solver solver)
-        {
-            Solver solver1 = solver.Clone();
-            Assert.AreEqual(0ul, solver1.CountSolutions(multiThread: true));
+    public static void TestInvalidSolution(this Solver solver)
+    {
+        Solver solver1 = solver.Clone();
+        Assert.AreEqual(0ul, solver1.CountSolutions(multiThread: true));
 
-            Solver solver2 = solver.Clone();
-            Assert.IsFalse(solver2.FindSolution(multiThread: true));
+        Solver solver2 = solver.Clone();
+        Assert.IsFalse(solver2.FindSolution(multiThread: true));
 
-            Solver solver3 = solver.Clone();
-            Assert.IsFalse(solver3.FillRealCandidates(multiThread: true));
-        }
+        Solver solver3 = solver.Clone();
+        Assert.IsFalse(solver3.FillRealCandidates(multiThread: true));
+    }
 
-        public static void TestMultipleSolution(this Solver solver)
-        {
-            Solver solver1 = solver.Clone();
-            Assert.IsTrue(solver1.CountSolutions(multiThread: true) > 1ul);
+    public static void TestMultipleSolution(this Solver solver)
+    {
+        Solver solver1 = solver.Clone();
+        Assert.IsTrue(solver1.CountSolutions(multiThread: true) > 1ul);
 
-            Solver solver2 = solver.Clone();
-            Assert.IsTrue(solver2.FindSolution(multiThread: true));
+        Solver solver2 = solver.Clone();
+        Assert.IsTrue(solver2.FindSolution(multiThread: true));
 
-            Solver solver3 = solver.Clone();
-            Assert.IsTrue(solver3.FillRealCandidates(multiThread: true));
-            Assert.IsFalse(solver3.IsComplete);
-        }
+        Solver solver3 = solver.Clone();
+        Assert.IsTrue(solver3.FillRealCandidates(multiThread: true));
+        Assert.IsFalse(solver3.IsComplete);
     }
 }
